+++ spring-cloud-starter-tencent-polaris-circuitbreaker/src/test/java/com/tencent/cloud/polaris/circuitbreaker/feign/PolarisFeignClientTest.java
@@ -18,6 +18,7 @@
 package com.tencent.cloud.polaris.circuitbreaker.feign;
 
 import com.tencent.cloud.polaris.circuitbreaker.PolarisFeignClientAutoConfiguration;
+import com.tencent.cloud.polaris.context.PolarisContextConfiguration;
 import feign.Client;
 import org.junit.Test;
 import org.junit.jupiter.api.Assertions;
@@ -36,7 +37,7 @@ import org.springframework.test.context.junit4.SpringRunner;
  */
 @RunWith(SpringRunner.class)
 @SpringBootTest(classes = TestPolarisFeignApp.class)
-@ContextConfiguration(classes = { PolarisFeignClientAutoConfiguration.class })
+@ContextConfiguration(classes = { PolarisFeignClientAutoConfiguration.class, PolarisContextConfiguration.class })
 public class PolarisFeignClientTest {
 
 	@Autowired