+++ spring-cloud-starter-tencent-polaris-discovery/src/test/java/com/tencent/cloud/polaris/discovery/PolarisServiceDiscoveryTest.java
@@ -53,7 +53,8 @@ public class PolarisServiceDiscoveryTest {
 			.withConfiguration(AutoConfigurations.of(PolarisContextConfiguration.class,
 					PolarisServiceDiscoveryTest.PolarisPropertiesConfiguration.class,
 					PolarisDiscoveryClientConfiguration.class,
-					PolarisDiscoveryAutoConfiguration.class))
+					PolarisDiscoveryAutoConfiguration.class,
+					PolarisContextConfiguration.class))
 			.withPropertyValues("spring.application.name=" + SERVICE_PROVIDER)
 			.withPropertyValues("server.port=" + PORT)
 			.withPropertyValues("spring.cloud.polaris.address=grpc://127.0.0.1:10081")